+++ adana-script/src/tests/drop.rs
