+++ adana-script/src/tests/mod.rs
@@ -1,20 +1,20 @@
+mod array;
 mod bitwise;
 mod builtin;
+mod chaining;
+mod drop;
 mod dynload;
 mod examples;
+mod file;
 mod foreach;
 mod funct;
 mod is_type;
 mod misc;
 mod opassign;
+mod parser;
 mod range;
+mod reference;
+mod scope_ctx;
 mod strings;
 mod struc;
-mod test_array;
-mod test_chaining;
-mod test_drop;
-mod test_parser;
-mod test_reference;
-mod test_scope_ctx;
-mod tests_file;
 mod unused;
