+++ adana-script/src/tests/scope_ctx.rs
